[#113] Reorganized folder structure to be more grouped by feature rather than type
