+++ packages/frontend/src/auth/auth.interface.ts
