+++ packages/frontend/src/pages/Login/index.test.tsx
@@ -3,7 +3,7 @@ import { MemoryRouter, Switch, Route } from "react-router-dom";
 import { render, fireEvent } from "@testing-library/react";
 import axios from "axios";
 import LoginPage from "./index";
-import { usernameErrors, passwordErrors } from "../../validations/user.schema";
+import { usernameErrors, passwordErrors } from "../../auth/auth.schema";
 
 jest.mock("axios");
 const axiosMock = axios as jest.Mocked<typeof axios>;
