+++ packages/frontend/src/pages/Login/index.tsx
@@ -2,9 +2,9 @@ import React, { useState } from "react";
 import { Redirect } from "react-router-dom";
 import { AxiosError } from "axios";
 import { useFormik } from "formik";
-import { userCredentialsSchema } from "../../validations/user.schema";
-import { login } from "../../services/auth.service";
-import { UserCredentials } from "../../interfaces/auth.interface";
+import { userCredentialsSchema } from "../../auth/auth.schema";
+import { login } from "../../auth/auth.service";
+import { UserCredentials } from "../../auth/auth.interface";
 import { ResponseError } from "../../interfaces/responseError.interface";
 import Heading from "../../components/Heading";
 import Text from "../../components/Text";
