+++ packages/frontend/src/pages/Recaps/components/Other/Recap/index.tsx
@@ -8,7 +8,7 @@ import Button from "../../../../../components/Button";
 import Icon from "../../../../../components/Icon";
 import BulletPoint from "../../BulletPoint";
 import RecapIcon from "../../RecapIcon";
-import { RecapOther } from "../../../../../interfaces/recaps.interface";
+import { RecapOther } from "../../../recaps.interface";
 import { CommonProps } from "../../../../../components/commonProps";
 
 interface OtherRecapProps extends CommonProps {
