+++ packages/frontend/src/pages/Recaps/components/RecapListCard/index.tsx
@@ -4,7 +4,7 @@ import Card from "../../../../components/Card";
 import Heading from "../../../../components/Heading";
 import Button from "../../../../components/Button";
 import RecapIcon, { RecapIconProps } from "../RecapIcon";
-import { RecapKind } from "../../../../interfaces/recaps.interface";
+import { RecapKind } from "../../recaps.interface";
 import { CommonProps } from "../../../../components/commonProps";
 
 interface IconProps extends CommonProps {
