+++ packages/frontend/src/pages/Recaps/components/SideProjects/Recap/index.tsx
@@ -8,7 +8,7 @@ import Button from "../../../../../components/Button";
 import Icon from "../../../../../components/Icon";
 import BulletPoint from "../../BulletPoint";
 import RecapIcon from "../../RecapIcon";
-import { RecapSideProjects } from "../../../../../interfaces/recaps.interface";
+import { RecapSideProjects } from "../../../recaps.interface";
 import { CommonProps } from "../../../../../components/commonProps";
 
 interface SideProjectsRecapProps extends CommonProps {
