+++ packages/frontend/src/pages/Recaps/components/Skills/Recap/index.tsx
@@ -7,7 +7,7 @@ import Button from "../../../../../components/Button";
 import Icon from "../../../../../components/Icon";
 import BulletPoint from "../../BulletPoint";
 import RecapIcon from "../../RecapIcon";
-import { RecapSkills } from "../../../../../interfaces/recaps.interface";
+import { RecapSkills } from "../../../recaps.interface";
 import { CommonProps } from "../../../../../components/commonProps";
 
 interface SkillsRecapProps extends CommonProps {
