+++ packages/frontend/src/pages/Signup/index.tsx
@@ -2,9 +2,9 @@ import React, { useState } from "react";
 import { Redirect } from "react-router-dom";
 import { AxiosError } from "axios";
 import { useFormik } from "formik";
-import { signupUserSchema } from "../../validations/user.schema";
-import { signup } from "../../services/auth.service";
-import { SignupUser } from "../../interfaces/auth.interface";
+import { signupUserSchema } from "../../auth/auth.schema";
+import { signup } from "../../auth/auth.service";
+import { SignupUser } from "../../auth/auth.interface";
 import { ResponseError } from "../../interfaces/responseError.interface";
 import Heading from "../../components/Heading";
 import Text from "../../components/Text";
