--- packages/frontend/src/services/defaultHeaders.ts
@@ -1,1 +1,0 @@
-// TODO: set up default headers with Authorization token
