--- packages/frontend/src/services/recaps.service.ts
@@ -1,7 +1,0 @@
-export const getAllRecaps = () => {};
-
-export const createRecap = () => {};
-
-export const updateRecap = () => {};
-
-export const deleteRecap = () => {};
